update toggle button styling, and move omit field stuff into a separate hook
